chore: move app to reconstruction dir
