+++ compositions/cluster-k8s/block-reconstruction/256-square-size/2-full-nodes.toml
@@ -1,214 +1,113 @@
 [metadata]
-  name = "reconstruction"
-  author = "Bidon15"
+name = "reconstruction"
+author = "Bidon15"
 
 [global]
-  plan = "celestia"
-  case = "reconstruction"
-  total_instances = 23099
-  concurrent_builds = 0
-  builder = "docker:generic"
-  runner = "cluster:k8s"
-  disable_metrics = false
-  [global.run]
-    artifact = ""
-    [global.run.test_params]
-      bridge = "377"
-      execution-time = "40"
-      full = "2"
-      light = "22615"
-      msg-size = "15000"
-      persistent-peers = "10"
-      seed = "5"
-      submit-times = "30"
-      validator = "100"
+plan = "celestia"
+case = "reconstruction"
+total_instances = 5852
+builder = "docker:generic"
+runner = "cluster:k8s"
+disable_metrics = false
+[global.run]
+artifact = ""
+[global.run.test_params]
+bridge = "100"
+execution-time = "40"
+full = "2"
+light = "5650"
+msg-size = "70000"
+persistent-peers = "10"
+submit-times = "40"
+validator = "100"
 
 [[groups]]
-  id = "seeds"
-  builder = "docker:generic"
-  [groups.build_config]
-    build_base_image = "golang:1.19.1"
-    enable_go_build_cache = true
-    enabled = true
-    go_version = "1.19"
-  [groups.build]
-  [groups.resources]
-    memory = "4Gi"
-    cpu = "3"
-  [groups.instances]
-    count = 5
-    percentage = 0.0
-  [groups.run]
-    artifact = ""
-    [groups.run.test_params]
-      bandwidth = "256Mib"
-      latency = "50"
-      role = "seed"
+id = "validators"
+builder = "docker:generic"
+[groups.resources]
+memory = "3Gi"
+cpu = "2000m"
+[groups.instances]
+count = 100
+percentage = 0.0
+[groups.build_config]
+build_base_image = "golang:1.20"
+enable_go_build_cache = true
+enabled = true
+go_version = "1.20"
+[groups.build]
+[groups.run]
+artifact = ""
+[groups.run.test_params]
+bandwidth = "256Mib"
+latency = "50"
+role = "validator"
 
 [[groups]]
-  id = "validators"
-  builder = "docker:generic"
-  [groups.build_config]
-    build_base_image = "golang:1.19.1"
-    enable_go_build_cache = true
-    enabled = true
-    go_version = "1.19"
-  [groups.build]
-  [groups.resources]
-    memory = "4Gi"
-    cpu = "4000m"
-  [groups.instances]
-    count = 100
-    percentage = 0.0
-  [groups.run]
-    artifact = ""
-    [groups.run.test_params]
-      bandwidth = "256Mib"
-      latency = "50"
-      role = "validator"
+id = "bridges"
+builder = "docker:generic"
+[groups.resources]
+memory = "3Gi"
+cpu = "2000m"
+[groups.instances]
+count = 100
+percentage = 0.0
+[groups.build_config]
+build_base_image = "golang:1.20"
+enable_go_build_cache = true
+enabled = true
+go_version = "1.20"
+[groups.build]
+[groups.run]
+artifact = ""
+[groups.run.test_params]
+bandwidth = "256Mib"
+block-height = "5"
+latency = "50"
+role = "bridge"
 
 [[groups]]
-  id = "bridges"
-  builder = "docker:generic"
-  [groups.build_config]
-    build_base_image = "golang:1.19.1"
-    enable_go_build_cache = true
-    enabled = true
-    go_version = "1.19"
-  [groups.build]
-  [groups.resources]
-    memory = "12Gi"
-    cpu = "5000m"
-  [groups.instances]
-    count = 377
-    percentage = 0.0
-  [groups.run]
-    artifact = ""
-    [groups.run.test_params]
-      bandwidth = "256Mib"
-      block-height = "5"
-      latency = "50"
-      role = "bridge"
+id = "fulls"
+builder = "docker:generic"
+[groups.resources]
+memory = "3Gi"
+cpu = "2000m"
+[groups.instances]
+count = 2
+percentage = 0.0
+[groups.build_config]
+build_base_image = "golang:1.20"
+enable_go_build_cache = true
+enabled = true
+go_version = "1.20"
+[groups.build]
+[groups.run]
+artifact = ""
+[groups.run.test_params]
+bandwidth = "256Mib"
+block-height = "5"
+latency = "50"
+role = "full"
 
 [[groups]]
-  id = "fulls"
-  builder = "docker:generic"
-  [groups.build_config]
-    build_base_image = "golang:1.19.1"
-    enable_go_build_cache = true
-    enabled = true
-    go_version = "1.19"
-  [groups.build]
-  [groups.resources]
-    memory = "24Gi"
-    cpu = "16"
-  [groups.instances]
-    count = 2
-    percentage = 0.0
-  [groups.run]
-    artifact = ""
-    [groups.run.test_params]
-      bandwidth = "256Mib"
-      block-height = "5"
-      latency = "50"
-      role = "full"
-
-[[groups]]
-  id = "lights"
-  builder = "docker:generic"
-  [groups.build_config]
-    build_base_image = "golang:1.19.1"
-    enable_go_build_cache = true
-    enabled = true
-    go_version = "1.19"
-  [groups.build]
-  [groups.resources]
-    memory = "400Mi"
-    cpu = "200m"
-  [groups.instances]
-    count = 22615
-    percentage = 0.0
-  [groups.run]
-    artifact = ""
-    [groups.run.test_params]
-      bandwidth = "256Mib"
-      block-height = "5"
-      latency = "50"
-      role = "light"
-
-[[runs]]
-  id = "default"
-  total_instances = 23099
-
-  [[runs.groups]]
-    id = "seeds"
-    group_id = "seeds"
-    [runs.groups.resources]
-      memory = "4Gi"
-      cpu = "3"
-    [runs.groups.instances]
-      count = 5
-      percentage = 0.0
-    [runs.groups.test_params]
-      bandwidth = "256Mib"
-      latency = "50"
-      role = "seed"
-
-  [[runs.groups]]
-    id = "validators"
-    group_id = "validators"
-    [runs.groups.resources]
-      memory = "4Gi"
-      cpu = "4000m"
-    [runs.groups.instances]
-      count = 100
-      percentage = 0.0
-    [runs.groups.test_params]
-      bandwidth = "256Mib"
-      latency = "50"
-      role = "validator"
-
-  [[runs.groups]]
-    id = "bridges"
-    group_id = "bridges"
-    [runs.groups.resources]
-      memory = "12Gi"
-      cpu = "5000m"
-    [runs.groups.instances]
-      count = 377
-      percentage = 0.0
-    [runs.groups.test_params]
-      bandwidth = "256Mib"
-      block-height = "5"
-      latency = "50"
-      role = "bridge"
-
-  [[runs.groups]]
-    id = "fulls"
-    group_id = "fulls"
-    [runs.groups.resources]
-      memory = "24Gi"
-      cpu = "16"
-    [runs.groups.instances]
-      count = 2
-      percentage = 0.0
-    [runs.groups.test_params]
-      bandwidth = "256Mib"
-      block-height = "5"
-      latency = "50"
-      role = "full"
-
-  [[runs.groups]]
-    id = "lights"
-    group_id = "lights"
-    [runs.groups.resources]
-      memory = "400Mi"
-      cpu = "200m"
-    [runs.groups.instances]
-      count = 22615
-      percentage = 0.0
-    [runs.groups.test_params]
-      bandwidth = "256Mib"
-      block-height = "5"
-      latency = "50"
-      role = "light"
+id = "lights"
+builder = "docker:generic"
+[groups.resources]
+memory = "300Mi"
+cpu = "150m"
+[groups.instances]
+count = 5650
+percentage = 0.0
+[groups.build_config]
+build_base_image = "golang:1.20"
+enable_go_build_cache = true
+enabled = true
+go_version = "1.20"
+[groups.build]
+[groups.run]
+artifact = ""
+[groups.run.test_params]
+bandwidth = "256Mib"
+block-height = "5"
+latency = "50"
+role = "light"
